Resolve issue with comparison rules and null values (+Pete).
Do not apply conditions for comparison rules if either the source or comparison value are null and the comparison type is not equality/inequality.  The thinking is that we can't accurately perform gt/lt comparisons with null, so if a comparison rule is to be in effect, it must have values to work with (except in the case of equality or inequality).  If requiredness is needed in addition to a comparison, then a separate required rule should be used.